requirements.txt
openai
manim
manim-voiceover
gTTS
python-dotenv
setuptools
opencv-python
numpy
pylint